// frontend/src/app/models/student.ts
export interface Student {
    id: any;
    studentName: string;
    fathersName: string;
    mothersName: string;
    mobile: string;
    altMobile: string;
    course: string;
    department: string;
    dob: string;
    gender: string;
    rollNo: number|null;
    email: string;
    pPin: string;
    pAddress: string;
    pCity: string;
    pState: string;
    cPin: string;
    cAddress: string;
    cCity: string;
    cState: string;
    category: string;
    maritalStatus: string;
    aadhaarNo: string;
    college: string;
    bloodGroup: string;
    admissionDate: Date|null;
    profilePhoto: string;
   
}